refactor v2 map step tests
